secur/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
discord.py==2.3.2
disutils==1.4.32.post2
docopt==0.6.2
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
idna==3.6
Js2Py==0.74
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
pipwin==0.5.2
PyAudio==0.2.14
pygame==2.5.2
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
ratelim==0.1.6
requests==2.31.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
speech==0.5.2
SpeechRecognition==3.10.1
twilio==8.10.3
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4
